resources/test/tests/functional/abortsignal.html

<!DOCTYPE html>
<title>Test#get_signal</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  "use strict";

  setup(() => {
    assert_implements_optional(window.AbortController, "No AbortController");
  });

  let signal;

  test(t => {
    signal = t.get_signal();
    assert_true(signal instanceof AbortSignal, "Returns an abort signal");
    assert_false(signal.aborted, "Signal should not be aborted before test end");
  }, "t.signal existence");

  test(t => {
    assert_true(signal.aborted, "Signal should be aborted after test end");
  }, "t.signal.aborted");
</script>
